Remove unused `something.cs` file and refactor monetary values in services and models to use `long` instead of `decimal` for better precision in currency handling. Update related methods and DTOs to reflect these changes across `UserService`, `RouletteService`, and `Bet` entities.

+++ VirtualRoulette/Models/DTOs/BetResponse.cs
@@ -8,5 +8,5 @@ public class BetResponse
    
     public int WinningNumber { get; set; }
     
-    public int WonAmountInCents { get; set; }
+    public long WonAmountInCents { get; set; }
 }
